Revert "Revert "move exclude/include container list to datadog.yaml" (#637)"
This reverts commit afe9d7348d5cac6096fd46d4362cffcd4cb02063.
