+++ pkg/collector/corechecks/containers/docker.go
@@ -29,15 +29,12 @@ const dockerCheckName = "docker"
 
 type dockerConfig struct {
 	//Url                    string             `yaml:"url"`
-	CollectContainerSize  bool     `yaml:"collect_container_size"`
-	CollectImagesStats    bool     `yaml:"collect_images_stats"`
-	CollectImageSize      bool     `yaml:"collect_image_size"`
-	Exclude               []string `yaml:"exclude"`
-	Include               []string `yaml:"include"`
-	ExcludePauseContainer bool     `yaml:"exclude_pause_container"`
-	Tags                  []string `yaml:"tags"`
-	CollectEvent          bool     `yaml:"collect_events"`
-	FilteredEventType     []string `yaml:"filtered_event_types"`
+	CollectContainerSize bool     `yaml:"collect_container_size"`
+	CollectImagesStats   bool     `yaml:"collect_images_stats"`
+	CollectImageSize     bool     `yaml:"collect_image_size"`
+	Tags                 []string `yaml:"tags"`
+	CollectEvent         bool     `yaml:"collect_events"`
+	FilteredEventType    []string `yaml:"filtered_event_types"`
 	//CustomCGroup           bool               `yaml:"custom_cgroups"`
 	//HealthServiceWhitelist []string           `yaml:"health_service_check_whitelist"`
 	//CollectContainerCount  bool               `yaml:"collect_container_count"`
@@ -53,9 +50,6 @@ type dockerConfig struct {
 
 const (
 	DockerServiceUp string = "docker.service_up"
-
-	pauseContainerGCR       string = "image:gcr.io/google_containers/pause.*"
-	pauseContainerOpenshift string = "image:openshift/origin-pod"
 )
 
 type containerPerImage struct {
@@ -70,10 +64,6 @@ func (c *dockerConfig) Parse(data []byte) error {
 	}
 	if len(c.FilteredEventType) == 0 {
 		c.FilteredEventType = []string{"top", "exec_create", "exec_start"}
-	}
-
-	if c.ExcludePauseContainer {
-		c.Exclude = append(c.Exclude, pauseContainerGCR, pauseContainerOpenshift)
 	}
 	return nil
 }
@@ -236,17 +226,9 @@ func (d *DockerCheck) String() string {
 func (d *DockerCheck) Configure(config, initConfig check.ConfigData) error {
 	d.instance = &dockerConfig{
 		// Default conf values
-		ExcludePauseContainer: true,
-		CollectEvent:          true,
+		CollectEvent: true,
 	}
 	d.instance.Parse(config)
-
-	docker.InitDockerUtil(&docker.Config{
-		CacheDuration:  10 * time.Second,
-		CollectNetwork: true,
-		Whitelist:      d.instance.Include,
-		Blacklist:      d.instance.Exclude,
-	})
 
 	var err error
 	d.dockerHostname, err = docker.GetHostname()
