+++ pkg/config/config.go
@@ -114,6 +114,7 @@ func init() {
 	Datadog.SetDefault("jmx_pipe_name", "dd-auto_discovery")
 	// Autoconfig
 	Datadog.SetDefault("autoconf_template_dir", "/datadog/check_configs")
+	Datadog.SetDefault("exclude_pause_container", true)
 	// Docker
 	Datadog.SetDefault("docker_labels_as_tags", map[string]string{})
 	Datadog.SetDefault("docker_env_as_tags", map[string]string{})
